+++ wasm/TextureGenerator.h
@@ -8,58 +8,58 @@ struct RGB {
 };
 
 struct XYZ {
-    double x, y, z;
+    float x, y, z;
 };
 
 class TextureGenerator {
    public:
     ~TextureGenerator();
 
-    double surfaceSeed = 0;
-    double landSeed = 1;
-    double cloudSeed = 2;
+    float surfaceSeed = 0;
+    float landSeed = 1;
+    float cloudSeed = 2;
 
     unsigned int resolution = 256;
 
-    double surfaceiScale = 2;
+    float surfaceiScale = 2;
     unsigned char surfaceiOctaves = 8;
-    double surfaceiFalloff = 1;
-    double surfaceiIntensity = 1;
-    double surfaceiRidginess = 0.5;
-    double surfacesScale = 1;
+    float surfaceiFalloff = 1;
+    float surfaceiIntensity = 1;
+    float surfaceiRidginess = 0.5;
+    float surfacesScale = 1;
     unsigned char surfacesOctaves = 0;
-    double surfacesFalloff = 1;
-    double surfacesIntensity = 1;
+    float surfacesFalloff = 1;
+    float surfacesIntensity = 1;
 
     RGB landColor1 = UL2RGB(0xe6af7e);
     RGB landColor2 = UL2RGB(0x007200);
-    double landiScale = 2;
+    float landiScale = 2;
     unsigned char landiOctaves = 1;
-    double landiFalloff = 1;
-    double landiIntensity = 1;
-    double landiRidginess = 0;
-    double landsScale = 2;
+    float landiFalloff = 1;
+    float landiIntensity = 1;
+    float landiRidginess = 0;
+    float landsScale = 2;
     unsigned char landsOctaves = 0;
-    double landsFalloff = 1;
-    double landsIntensity = 1;
+    float landsFalloff = 1;
+    float landsIntensity = 1;
 
     RGB waterDeep = UL2RGB(0x000033);
     RGB waterShallow = UL2RGB(0x0000ff);
-    double waterLevel = 0.68;
-    double waterSpecular = 1;
-    double waterFalloff = 1;
+    float waterLevel = 0.68;
+    float waterSpecular = 1;
+    float waterFalloff = 1;
 
     RGB cloudColor = UL2RGB(0xffffff);
-    double cloudOpacity = 0.75;
-    double cloudiScale = 0.5;
+    float cloudOpacity = 0.75;
+    float cloudiScale = 0.5;
     unsigned char cloudiOctaves = 2;
-    double cloudiFalloff = 2;
-    double cloudiIntensity = 1.8;
-    double cloudiRidginess = 0;
-    double cloudsScale = 0.5;
+    float cloudiFalloff = 2;
+    float cloudiIntensity = 1.8;
+    float cloudiRidginess = 0;
+    float cloudsScale = 0.5;
     unsigned char cloudsOctaves = 5;
-    double cloudsFalloff = 1;
-    double cloudsIntensity = 1;
+    float cloudsFalloff = 1;
+    float cloudsIntensity = 1;
 
     unsigned char *diffuseBuffer;
     unsigned char *normalBuffer;
@@ -71,7 +71,7 @@ class TextureGenerator {
     NoiseWrapper *cloudNoise;
 
     void init();
-    RGB surfaceColor(double x, double y, double z);
+    RGB surfaceColor(float x, float y, float z);
     RGB UL2RGB(unsigned long dwColor);
     void setPixel(unsigned char *buffer, unsigned int x, unsigned int y,
                   RGB color);
